iommu/samsung: Consolidate log message about gate clk
Integrate the log message about the non-existence of a gate clock into
the one log message that we print on successful probe. This helps with
reducing the amount of log messages on boot up.

This is an example of the combined log message:

samsung-sysmmu 1c120000.sysmmu: initialized IOMMU. Ver 8.1.0 no gate clock

Bug: 178756420
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: If78a95782311874bd5a0028b35cbf928c955f371